# Optimize `Deque#concat(Indexable)`